Migration scripts from modrepo, data source improvements / fixes, upload service created

+++ grails-app/domain/org/openmrs/modulus/Uploadable.groovy
@@ -12,22 +12,20 @@ class Uploadable {
 
     String filename
     String contentType
-    byte[] rawFile
     String downloadURL
 
     String path
 
 
     static constraints = {
-        rawFile maxSize: 10 * 1024 * 1024, nullable: true
         path nullable: true
-        downloadURL maxLength: 255, nullable: true
-        filename maxLength: 255, nullable: true
+        downloadURL maxSize: 255, nullable: true
+        filename maxSize: 255, nullable: true
         contentType nullable: true
     }
 
     def beforeUpdate() {
-        if (isDirty('path')) {
+        if (isDirty('path') && this.getPersistentValue('path')) {
             // delete the old file
             Path oldPath = Paths.get(this.getPersistentValue('path'))
             Files.deleteIfExists(oldPath)
